Rename to version 0.0.92
