+++ upgrades/0.0.92.rotate-argocd-ssh-key/pkg/github/github.go
