+++ upgrades/0.0.92.rotate-argocd-ssh-key/pkg/okctlinit/cluster_declaration.go
