+++ upgrades/0.0.92.rotate-argocd-ssh-key/upgrade.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	"fmt"
-	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.89.rotate-argocd-ssh-key/pkg/cmdflags"
-	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.89.rotate-argocd-ssh-key/pkg/rotatesshkey"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.92.rotate-argocd-ssh-key/pkg/cmdflags"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.92.rotate-argocd-ssh-key/pkg/rotatesshkey"
 )
 
 func upgrade(context Context, flags cmdflags.Flags) error {
